--- app/(dashboard)/dashboard.css
@@ -1,65 +1,0 @@
-:root {
-  --color-primary: 110 242 255; /* #6EF2FF */
-  --color-secondary: 250 207 6; /* #FACF06 */
-}
-
-.eq-bar {
-  transform: scale(1, -1) translate(0, -24px);
-}
-
-.eq-bar--1 {
-  animation-name: bar-1;
-  animation-duration: 0.8s;
-  animation-iteration-count: infinite;
-}
-
-.eq-bar--2 {
-  animation-name: bar-2;
-  animation-duration: 0.8s;
-  animation-iteration-count: infinite;
-}
-
-.eq-bar--3 {
-  animation-name: bar-3;
-  animation-duration: 0.8s;
-  animation-iteration-count: infinite;
-}
-
-@keyframes bar-1 {
-  0% {
-    height: 5px;
-  }
-  50% {
-    height: 15px;
-  }
-  100% {
-    height: 5px;
-  }
-}
-
-@keyframes bar-2 {
-  0% {
-    height: 15px;
-  }
-  50% {
-    height: 5px;
-  }
-  100% {
-    height: 15px;
-  }
-}
-
-@keyframes bar-3 {
-  0% {
-    height: 10px;
-  }
-  40% {
-    height: 5px;
-  }
-  60% {
-    height: 15px;
-  }
-  100% {
-    height: 10px;
-  }
-}
