--- app/(site)/page.tsx
@@ -1,5 +1,0 @@
-const HomePage = () => {
-  return <div>Coming soon...</div>;
-};
-
-export default HomePage;
