+++ app/globals.css
@@ -2,6 +2,13 @@
 @tailwind components;
 @tailwind utilities;
 
+@layer base {
+  :root {
+    --color-primary: 110 242 255; /* #6EF2FF */
+    --color-secondary: 250 207 6; /* #FACF06 */
+  }
+}
+
 @layer components {
   #content {
     /* px-4 */
@@ -24,6 +31,69 @@
       /* px-8 */
       padding-left: max(2rem, env(safe-area-inset-left));
       padding-right: max(2rem, env(safe-area-inset-right));
+    }
+  }
+
+  .eq-bar {
+    transform: scale(1, -1) translate(0, -24px);
+  }
+
+  .eq-bar--1 {
+    animation-name: bar-1;
+    animation-duration: 0.8s;
+    animation-iteration-count: infinite;
+  }
+
+  .eq-bar--2 {
+    animation-name: bar-2;
+    animation-duration: 0.8s;
+    animation-iteration-count: infinite;
+  }
+
+  .eq-bar--3 {
+    animation-name: bar-3;
+    animation-duration: 0.8s;
+    animation-iteration-count: infinite;
+  }
+}
+
+@layer utilities {
+  @keyframes bar-1 {
+    0% {
+      height: 5px;
+    }
+    50% {
+      height: 15px;
+    }
+    100% {
+      height: 5px;
+    }
+  }
+
+  @keyframes bar-2 {
+    0% {
+      height: 15px;
+    }
+    50% {
+      height: 5px;
+    }
+    100% {
+      height: 15px;
+    }
+  }
+
+  @keyframes bar-3 {
+    0% {
+      height: 10px;
+    }
+    40% {
+      height: 5px;
+    }
+    60% {
+      height: 15px;
+    }
+    100% {
+      height: 10px;
     }
   }
 }
