+++ app/layout.tsx
@@ -1,14 +1,14 @@
 import "./globals.css";
 
 import { Metadata } from "next";
+import { Oxanium } from "next/font/google";
 
-import { siteConfig } from "@/lib/utils/site-config";
+const oxanium = Oxanium({
+  subsets: ["latin"],
+  variable: "--font-oxanium",
+});
 
 export const metadata: Metadata = {
-  title: {
-    default: siteConfig.title,
-    template: `%s ${siteConfig.titleSeparator} ${siteConfig.title}`,
-  },
   viewport: {
     width: "device-width",
     initialScale: 1,
@@ -25,7 +25,11 @@ export const metadata: Metadata = {
 };
 
 const RootLayout = ({ children }: { children: React.ReactNode }) => {
-  return <>{children}</>;
+  return (
+    <html lang="en" className={oxanium.variable}>
+      <body className="bg-gray-950 text-gray-400">{children}</body>
+    </html>
+  );
 };
 
 export default RootLayout;
