+++ components/Statistics.tsx
