+++ components/WeatherChart.tsx
